Split MigrationWorkflow tests into separate test class
Move the 9 tests that target MigrationWorkflow (GetMigrationCandidates, BuildMigrationSelections, MergeInstallRequests) out of InitWorkflowTests.cs and into a new MigrationWorkflowTests.cs file, mirroring the recent split of MigrationWorkflow from InitWorkflows.
